Finalized tutorial, puzzles and dialogue
This commit includes major changes for the Gate 2 submission point with a complete tutorial prototype, access to the main map and its inventory, diary and character customization menus. The dialogue system was also sort of finalized and finally we included our game design document to the project for people who are interested.