Tweaks to dialogue, parallax (in progress), camera...
Also made the interfaces such as settings and diary and inventory close when you tap anywhere but in the window on the screen. Added sound effects to the dialogue and highlighting to indicate who is currently speaking. Also fixed the camera not behaving properly in the escape game.

Hotfixed parallax to not end